Add fast iteration mode and mixed precision support for training pipeline
- Introduced `--fast` flag in `run_pipeline.sh` for quick training runs with reduced dataset and epochs.
- Updated `get_sequence_paths` to support limiting the number of sequence files for faster experiments.
- Enhanced training script to accept mixed precision flag and adjusted model output layer to maintain float32 precision.
- Added tests for max files truncation and mixed precision behavior in dataset and model tests.

+++ src/model.py
@@ -52,7 +52,8 @@ def build_model(
 
     x = tf.keras.layers.LSTM(lstm_units, return_sequences=False, name=f"lstm_{num_lstm_layers}")(x)
     x = tf.keras.layers.Dropout(dropout_rate, name="dropout")(x)
-    outputs = tf.keras.layers.Dense(vocab_size, activation="softmax", name="output")(x)
+    # dtype='float32' keeps logits/softmax in full precision under mixed_float16 policy.
+    outputs = tf.keras.layers.Dense(vocab_size, activation="softmax", dtype="float32", name="output")(x)
 
     model = tf.keras.Model(inputs=inputs, outputs=outputs, name="lstm_music_generator")
     _log_summary(model)
